Streamlit/requirements.txt
duckdb==0.8.0
python-dotenv==1.0.0
pandas==2.0.2
streamlit==1.23.1
